requirements.txt
temporalio
haystack-ai
uvicorn
fastapi